requirements.txt
Babel==2.5.1
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11
-e git+http://github.com/fusionbox/django-authtools@7842c4d52e0a25fd473fbd35db9d466259c0492a#egg=django_authtools
django-countries==5.0
-e git+https://github.com/mjs7231/django-dbbackup.git@7f3cbb214e9c4012ca267751c6560d6b7f4e7c5c#egg=django_dbbackup
django-language-field==0.0.3
django-phonenumber-field==1.3.0
django-rosetta==0.7.14
django-storages==1.6.5
django-tinymce4-lite==1.7.0
django-widget-tweaks==1.4.1
djangodeletes==0.43
dropbox==8.5.0
googletrans==2.2.0
gunicorn==19.7.1
idna==2.6
jsmin==2.2.2
microsofttranslator==0.8
phonenumberslite==8.8.8
Pillow==5.0.0
polib==1.1.0
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
